requirements.txt
Flask
pycryptodome
Flask==2.2.5

